Split out some helper functions from record socket
