+++ qcfractal/qcfractal/components/optimization/test_record_socket.py
@@ -21,7 +21,7 @@ from qcportal.singlepoint import (
     SinglepointProtocols,
 )
 from qcportal.utils import now_at_utc
-from ..record_socket import build_extras_properties
+from ..record_utils import build_extras_properties
 
 if TYPE_CHECKING:
     from qcfractal.db_socket import SQLAlchemySocket
